requirements.txt
anthropic
matplotlib
pandas
pytest
python-dotenv
streamlit
